# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server
components:
  securitySchemes:
    ApiKeyAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    Hippo:
      type: object
      properties:
        name:
          type: string
        weightKg:
          type: number
        color:
          type: string
        habitat:
          type: string
        maxSpeedKmHr:
          type: number
        birthDate:
          type: string
          format: date
        createdBy:
          type: string
        lastUpdatedBy:
          type: string

    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        date:
          type: string

paths:
  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string

      responses:
        200:
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string

  /hippos:
    get:
      summary: Retrieves a list of hippos
      description: Retrieves a list of hippos as JSON objects.
      responses:
        200:
          description: A list of Hippo JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hippo"

    post:
      summary: Create a new Hippo
      description: Create a new Hippo
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Hippo"

      responses:
        201:
          description: Hippo created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hippo"

  /hippos/fetch-sorted:
    post:
      summary: Retrieves a sorted and paginated list of hippos
      description: Retrieves a sorted and paginated list of hippos as JSON objects.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sortField:
                  type: string
                  description: The field name to sort by.
                  example: "name"
                sortOrder:
                  type: string
                  description: The order of sorting, can be 'asc' or 'desc'.
                  example: "asc"
                page:
                  type: integer
                  description: The page number to fetch.
                  example: 1
                limit:
                  type: integer
                  description: The number of items per page.
                  example: 10
              required:
                - page
                - limit
      responses:
        "200":
          description: A list of hippos, sorted and paginated.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hippo"

  /hippos/{id}:
    get:
      summary: Specific Hippo
      description: Retrieves a specific Hippo based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A Hippo in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hippo"

    put:
      summary: Updates a specific Hippo
      description: Updates a specific Hippo based on it id
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Hippo"

      responses:
        201:
          description: Hippo updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hippo"

    delete:
      summary: Deletes a specific Hippo
      description: Deletes a specific Hippo based on it id
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string

      responses:
        201:
          description: Hippo deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hippo"